requirements.txt
fastapi ==0.97.0
pydantic==1.10.9
uvicorn==0.22.0
httpx==0.24.0
python-dotenv==0.19.0